refactor(gateway): move RequestRetryHandler back to gateway
It turns out the `RequestRetryHandler` is actually a consumer of the
broker client API, building a specific retry loop around the broker
client's capabilities to enable sending CreateProcessInstance commands
to different partitions on specific errors instead of always retrying
the same partition (what the broker client will normally do).
