+++ gateway/src/main/java/io/camunda/zeebe/gateway/EndpointManager.java
@@ -14,9 +14,9 @@ import io.camunda.zeebe.broker.client.api.BrokerClient;
 import io.camunda.zeebe.broker.client.api.BrokerClusterState;
 import io.camunda.zeebe.broker.client.api.BrokerTopologyManager;
 import io.camunda.zeebe.broker.client.api.dto.BrokerRequest;
-import io.camunda.zeebe.broker.client.impl.RequestRetryHandler;
 import io.camunda.zeebe.gateway.ResponseMapper.BrokerResponseMapper;
 import io.camunda.zeebe.gateway.grpc.ServerStreamObserver;
+import io.camunda.zeebe.gateway.impl.broker.RequestRetryHandler;
 import io.camunda.zeebe.gateway.impl.broker.request.BrokerActivateJobsRequest;
 import io.camunda.zeebe.gateway.impl.configuration.MultiTenancyCfg;
 import io.camunda.zeebe.gateway.impl.job.ActivateJobsHandler;
